refactor(TimeService): 코드 리팩토링
- 시간 삭제 로직 리팩토링

+++ src/main/java/roomescape/time/service/TimeService.java
@@ -1,17 +1,18 @@
 package roomescape.time.service;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.global.exception.ErrorCode;
 import roomescape.global.exception.RoomescapeException;
-import roomescape.reservation.Reservation;
 import roomescape.reservation.dao.ReservationDao;
 import roomescape.time.AvailableTime;
 import roomescape.time.ReservationTime;
 import roomescape.time.dao.TimeDao;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+
+import static roomescape.global.exception.ErrorCode.CANNOT_DELETE_RESERVED_TIME;
 
 @Service
 public class TimeService {
@@ -44,16 +45,9 @@ public class TimeService {
 
     @Transactional
     public void deleteById(Long id) {
-        List<Reservation> reservations = reservationDao.selectByTimeId(id);
-        for (Reservation reservation : reservations) {
-            validateExistTimeByReservation(reservation.getDate(), reservation.getTime(), id);
+        if (reservationDao.existsUpcomingByTimeId(id, LocalDate.now(), LocalTime.now())) {
+            throw new RoomescapeException(CANNOT_DELETE_RESERVED_TIME);
         }
         timeDao.deleteById(id);
-    }
-
-    private void validateExistTimeByReservation(LocalDate date, ReservationTime time, Long timeId) {
-        if (!time.isBeforeDateTime(date, time) && time.getId().equals(timeId)) {
-            throw new RoomescapeException(ErrorCode.CANNOT_DELETE_RESERVED_TIME);
-        }
     }
 }
